requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
appnope==0.1.2
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.1.0
    # via notebook
attrs==21.2.0
    # via
    #   fiona
    #   jsonschema
backcall==0.2.0
    # via ipython
bleach==4.1.0
    # via nbconvert
certifi==2021.10.8
    # via
    #   fiona
    #   pyproj
cffi==1.15.0
    # via argon2-cffi
click==8.0.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
colorlover==0.3.0
    # via cufflinks
cufflinks==0.17.3
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
datetime==4.3
    # via -r requirements.in
debugpy==1.5.1
    # via ipykernel
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via
    #   jupyter-client
    #   nbconvert
fastdtw==0.3.4
    # via -r requirements.in
fiona==1.8.20
    # via geopandas
fonttools==4.28.3
    # via matplotlib
geopandas==0.10.2
    # via
    #   -r requirements.in
    #   keplergl
importlib-resources==5.4.0
    # via jsonschema
ipykernel==6.6.0
    # via
    #   ipywidgets
    #   notebook
ipython==7.30.1
    # via
    #   cufflinks
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via
    #   cufflinks
    #   keplergl
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   nbconvert
    #   notebook
jsonschema==4.2.1
    # via nbformat
jupyter-client==7.1.0
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
keplergl==0.3.2
    # via -r requirements.in
kiwisolver==1.3.2
    # via matplotlib
llvmlite==0.37.0
    # via numba
markupsafe==2.0.1
    # via jinja2
matplotlib==3.5.0
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
munch==2.5.0
    # via fiona
nbclient==0.5.9
    # via nbconvert
nbconvert==6.3.0
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.6
    # via widgetsnbextension
numba==0.54.1
    # via stumpy
numpy==1.20.3
    # via
    #   -r requirements.in
    #   cufflinks
    #   fastdtw
    #   matplotlib
    #   numba
    #   pandas
    #   scipy
    #   seaborn
    #   stumpy
packaging==21.3
    # via
    #   bleach
    #   matplotlib
    #   setuptools-scm
pandas==1.3.4
    # via
    #   -r requirements.in
    #   cufflinks
    #   geopandas
    #   keplergl
    #   seaborn
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.4.0
    # via matplotlib
plotly==5.7.0
    # via
    #   -r requirements.in
    #   cufflinks
prometheus-client==0.12.0
    # via notebook
prompt-toolkit==3.0.23
    # via ipython
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pycparser==2.21
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
pyproj==3.3.0
    # via geopandas
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2021.3
    # via
    #   datetime
    #   pandas
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
scipy==1.7.3
    # via
    #   seaborn
    #   stumpy
seaborn==0.11.2
    # via -r requirements.in
send2trash==1.8.0
    # via notebook
setuptools-scm==6.3.2
    # via matplotlib
shapely==1.8.0
    # via
    #   -r requirements.in
    #   geopandas
    #   keplergl
six==1.16.0
    # via
    #   bleach
    #   cufflinks
    #   fiona
    #   munch
    #   plotly
    #   python-dateutil
stumpy==1.10.0
    # via -r requirements.in
tenacity==8.0.1
    # via plotly
terminado==0.12.1
    # via notebook
testpath==0.5.0
    # via nbconvert
tomli==1.2.2
    # via setuptools-scm
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   traittypes
traittypes==0.2.1
    # via keplergl
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
widgetsnbextension==3.5.2
    # via ipywidgets
zipp==3.6.0
    # via importlib-resources
zope.interface==5.4.0
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools

libpysal~=4.6.2
esda~=2.4.1
scikit-learn~=1.0.2
tslearn~=0.5.2